# src/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "User, Project, and View Management API",
      "description": "API documentation for managing users, projects, and views."
    },
    "host": "localhost:5000",
    "basePath": "/api",
    "schemes": ["http"],
    "paths": {
      "/users": {
        "get": {
          "summary": "Get all users",
          "description": "Fetch all users with their associated groups.",
          "responses": {
            "200": {
              "description": "List of users",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        },
        "post": {
          "summary": "Create a new user",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/NewUser"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "User created successfully",
              "schema": {
                "$ref": "#/definitions/User"
              }
            },
            "400": {
              "description": "Validation error"
            },
            "409": {
              "description": "User already exists"
            }
          }
        }
      },
      "/users/{id}": {
        "get": {
          "summary": "Get user by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string",
              "description": "User ID"
            }
          ],
          "responses": {
            "200": {
              "description": "User details",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/definitions/User"
                  }
                }
              }
            },
            "404": {
              "description": "User not found"
            },
            "500": {
              "description": "Unexpected error"
            }
          }
        },
        "delete": {
          "summary": "Delete user by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string",
              "description": "User ID"
            }
          ],
          "responses": {
            "204": {
              "description": "User deleted successfully"
            },
            "404": {
              "description": "User not found"
            }
          }
        }
      },
      "/users/login": {
        "post": {
          "summary": "Log in user",
          "parameters": [
        {
          "in": "body",
          "name": "body",
          "required": true,
          "schema": {
            "$ref": "#/definitions/LoginUser"
          }
        }
          ],
          "responses": {
        "200": {
          "description": "User logged in successfully"
        },
        "401": {
          "description": "Unauthorized"
        }
          }
        }
      },
      "/users/logout": {
        "post": {
          "summary": "Log out user",
          "responses": {
            "200": {
              "description": "User logged out successfully"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        }
      },
      "/users/groups": {
        "get": {
          "summary": "Get all groups",
          "description": "Fetch all groups.",
          "responses": {
            "200": {
              "description": "List of groups",
              "schema": {
                "$ref": "#/definitions/Group"
              }
            }
          }
        }
        ,


        "post": {
          "summary": "Create a new group",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/NewGroup"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Group created successfully",
              "schema": {
                "$ref": "#/definitions/Group"
              }
            },
            "409": {
              "description": "Group already exists"
            }
          }
        }
      },
      "/users/groups/{id}": {
        "delete": {
          "summary": "Delete group by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string",
              "description": "Group ID"
            }
          ],
          "responses": {
            "204": {
              "description": "Group deleted successfully"
            },
            "404": {
              "description": "Group not found"
            }
          }
        }
      },
...
    },
    "definitions": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "group": {
            "type": "string"
          }
        }
      },
      "NewUser": {
        "type": "object",
        "required": ["username", "email", "name", "password", "group"],
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "group": {
            "type": "string"
          }
        }
      },
      "LoginUser": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
        "type": "string"
          },
          "password": {
        "type": "string"
          }
        }
      },
      "Group": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "NewGroup": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  }